+++ example/lib/main.dart
@@ -20,8 +20,23 @@ class MyApp extends StatelessWidget {
   }
 }
 
-class MyHomePage extends StatelessWidget {
+class MyHomePage extends StatefulWidget {
   const MyHomePage({super.key});
+
+  @override
+  State<MyHomePage> createState() => _MyHomePageState();
+}
+
+class _MyHomePageState extends State<MyHomePage> {
+  DateTime? _date;
+
+  String display() {
+    if (_date == null) {
+      return 'NONE';
+    } else {
+      return 'year:${_date!.year}\nmonth:${_date!.month}\nday:${_date!.day}';
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -35,16 +50,29 @@ class MyHomePage extends StatelessWidget {
           padding: const EdgeInsets.symmetric(horizontal: 20),
           child: Column(
             crossAxisAlignment: CrossAxisAlignment.start,
-            children: const [
-              SizedBox(height: 20),
-              Text('Date Formater field'),
+            children: [
+              const SizedBox(height: 20),
+              Text(display()),
+              const SizedBox(height: 20),
+              const Text('Date Formater field'),
               DateFormatField(
                 type: DateFormatType.type4,
-                decoration: InputDecoration(
-                    hintText: 'DD/MM/YY',
-                    border: InputBorder.none,
-                    icon: Icon(Icons.calendar_month)),
-              )
+                addCalendar: false,
+                decoration: const InputDecoration(
+                  labelStyle: TextStyle(
+                    fontWeight: FontWeight.bold,
+                    fontSize: 18,
+                    fontStyle: FontStyle.italic,
+                  ),
+                  // border: InputBorder.none,
+                  label: Text("Date"),
+                ),
+                onComplete: (date) {
+                  setState(() {
+                    _date = date;
+                  });
+                },
+              ),
             ],
           ),
         ),