+++ lib/date_format_field.dart
@@ -2,3 +2,5 @@ library date_format_field;
 
 export 'src/date_field.dart';
 export 'src/enums.dart' show FormaterType;
+
+/// [DateFormatField] provides an auto format option to textfield