+++ lib/src/date_field.dart
@@ -12,15 +12,29 @@ enum DateFormatType {
 class DateFormatField extends StatefulWidget {
   const DateFormatField({
     super.key,
+    required this.onComplete,
     required this.type,
+    this.addCalendar = true,
     this.decoration,
   });
 
-  /// textfield decoration
+  /// [InputDecoration] a styling class for form field
+  ///
+  /// This is the default flutter Input decoration used to style input fields
   final InputDecoration? decoration;
 
-  /// date format type
+  /// [DateFormatType] is an enum for specifying the type
   final DateFormatType type;
+
+  /// [onSubmit] returns a nullable Datetime object
+  ///
+  /// Returns null when the datetime field is not complete
+  /// Returns a datetime object when the field has been completed
+  final Function(DateTime?) onComplete;
+
+  /// [addCalendar] sets a button that allows the selection of date from a
+  /// calendar pop up
+  final bool addCalendar;
 
   @override
   State<DateFormatField> createState() => _DateFormatFieldState();
@@ -29,24 +43,49 @@ class DateFormatField extends StatefulWidget {
 class _DateFormatFieldState extends State<DateFormatField> {
   final TextEditingController _dobFormater = TextEditingController();
 
+  InputDecoration? decoration() {
+    if (!widget.addCalendar) {
+      return widget.decoration;
+    }
+    if (widget.decoration == null) {
+      return InputDecoration(
+          suffixIcon: IconButton(
+              onPressed: () {
+                pickDate();
+              },
+              icon: Icon(Icons.calendar_month)));
+    }
+    return widget.decoration!.copyWith(
+      suffixIcon: IconButton(
+          onPressed: () {
+            pickDate();
+          },
+          icon: Icon(Icons.calendar_month)),
+    );
+  }
+
   void formatInput(String value) {
     /// formater for the text input field
+    DateTime? _completeDate;
     switch (widget.type) {
       case DateFormatType.type1:
-        Formater.type1(value, _dobFormater);
+        _completeDate = Formater.type1(value, _dobFormater);
         break;
       case DateFormatType.type2:
-        Formater.type2(value, _dobFormater);
+        _completeDate = Formater.type2(value, _dobFormater);
         break;
       case DateFormatType.type3:
-        Formater.type3(value, _dobFormater);
+        _completeDate = Formater.type3(value, _dobFormater);
         break;
       case DateFormatType.type4:
-        Formater.type4(value, _dobFormater);
+        _completeDate = Formater.type4(value, _dobFormater);
         break;
       default:
     }
-    setState(() {});
+    setState(() {
+      // update the datetime
+      widget.onComplete(_completeDate);
+    });
   }
 
   Future<void> pickDate() async {
@@ -69,6 +108,12 @@ class _DateFormatFieldState extends State<DateFormatField> {
   Widget build(BuildContext context) {
     return TextField(
       controller: _dobFormater,
+      onTap: () {
+        _dobFormater.selection = TextSelection.fromPosition(
+          TextPosition(offset: _dobFormater.text.length),
+        );
+      },
+      decoration: decoration(),
       keyboardType: TextInputType.datetime,
       onChanged: formatInput,
     );